# api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  handler.acceptNegotiationRequest:
    properties:
      acceptor_account_id:
        type: integer
      on_behalf_of_fund_id:
        description: |-
          OnBehalfOfFundID, when non-zero, places this accept on behalf of a fund (E2).
          The acceptor_account_id must equal the fund's RSD account.
          Caller must be the fund's manager (acting_employee_id enforced in stock-service).
        type: integer
    type: object
...
  handler.counterNegotiationRequest:
    properties:
      premium:
        type: string
      quantity:
        type: string
      settlement_date:
        type: string
      strike_price:
        type: string
    type: object
...
  handler.openNegotiationRequest:
    properties:
      bidder_account_id:
        type: integer
      premium:
        type: string
      quantity:
        type: string
      settlement_date:
        type: string
      strike_price:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: EXBanka Banking Microservices API Gateway. All endpoints are served
    under /api/v3 — v1 and v2 have been retired (plan E, 2026-04-27). Future versions
    (v4+) will be added as separate explicit router files with no transparent fallback.
    See api-gateway/internal/router/router_versioning.md.
  title: EXBanka API
  version: "3.0"
paths:
...
  /api/v3/me/otc/options/{id}:
    delete:
      description: Initiator-only. Status flips to "cancelled" and all open child
        negotiation chains cascade-cancel. Returns 204 on success, 403 if caller is
        not the listing's poster, 409 if listing is not open.
      parameters:
      - description: parent listing id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel (withdraw) your own OTC option listing
      tags:
      - OTCOptions
  /api/v3/me/otc/options/{id}/negotiations/{nid}:
    delete:
      description: Bidder-only — the listing's poster cannot cancel a bidder's chain
        (use reject instead).
      parameters:
      - description: parent listing id (sanity check)
        in: path
        name: id
        required: true
        type: integer
      - description: negotiation chain id
        in: path
        name: nid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Cancel (withdraw) your own OTC option negotiation chain
      tags:
      - OTCOptions
  /api/v3/me/otc/options/{id}/negotiations/{nid}/accept:
    post:
      consumes:
      - application/json
      description: 'Caller must be the party OPPOSITE to whoever proposed the current
        terms. After the negotiation state TX (which flips this chain to "accepted",
        parent to "consumed", and cascade-cancels every sibling chain), the contract-formation
        saga runs: mints OptionContract from the negotiation''s snapshot terms, reserves
        seller''s underlying shares, reserves+settles buyer''s premium, credits the
        seller. If the saga fails (e.g. seller no longer has the shares, buyer is
        short on premium), the negotiation flips to "failed" and the parent stays
        consumed.'
      parameters:
      - description: parent listing id
        in: path
        name: id
        required: true
        type: integer
      - description: negotiation chain id
        in: path
        name: nid
        required: true
        type: integer
      - description: acceptor_account_id — caller's account that pays the premium
          (if accepter is the buyer) or receives it (if accepter is the seller)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.acceptNegotiationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: acceptor_account_id missing or not owned
          schema:
            additionalProperties: true
            type: object
        "403":
          description: caller proposed current terms or not a party
          schema:
            additionalProperties: true
            type: object
        "409":
          description: parent listing already consumed
          schema:
            additionalProperties: true
            type: object
        "412":
          description: contract-formation saga rejected (seller short on shares OR
            buyer short on premium)
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Accept the current terms on an OTC option negotiation chain
      tags:
      - OTCOptions
  /api/v3/me/otc/options/{id}/negotiations/{nid}/counter:
    post:
      consumes:
      - application/json
      description: Either the bidder or the listing's poster may counter. Updates
        the chain's terms; status flips to "countered".
      parameters:
      - description: parent listing id (sanity check; not used)
        in: path
        name: id
        required: true
        type: integer
      - description: negotiation chain id
        in: path
        name: nid
        required: true
        type: integer
      - description: new terms
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.counterNegotiationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Counter on one of your OTC option negotiation chains
      tags:
      - OTCOptions
  /api/v3/me/otc/options/{id}/negotiations/{nid}/reject:
    post:
      description: Either party may reject. The chain ends without forming a contract;
        parent listing stays open.
      parameters:
      - description: parent listing id (sanity check)
        in: path
        name: id
        required: true
        type: integer
      - description: negotiation chain id
        in: path
        name: nid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Reject an OTC option negotiation chain
      tags:
      - OTCOptions
  /api/v3/me/otc/options/negotiations:
    get:
      description: Returns a unified list of the caller's LOCAL (intra-bank bidder)
        and REMOTE (cross-bank peer) negotiation chains. Each item carries `kind`
        (local|remote), `routing_number`/`bank_code` provenance, and `me_owner` (true
        only when the caller is the parent listing's poster/seller). Filter with `?statuses=open,countered,accepted,rejected,cancelled,expired`
        (applied to both sets).
      parameters:
      - description: comma-separated; omit for all
        in: query
        name: statuses
        type: string
      - description: 1-based, default 1
        in: query
        name: page
        type: integer
      - description: default 20, max 200
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List the caller's OTC option negotiation chains (local + remote, merged)
      tags:
      - OTCOptions
  /api/v3/me/otc/options/negotiations/{nid}/revisions:
    get:
      description: Returns all bid/counter/accept/reject revisions for a negotiation
        chain in revision_number order. Caller must be either the bidder or the parent
        listing's poster.
      parameters:
      - description: negotiation chain id
        in: path
        name: nid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: caller is not a party to this negotiation
          schema:
            additionalProperties: true
            type: object
        "404":
          description: negotiation not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List the full revision chain of an OTC option negotiation
      tags:
      - OTCOptions
...
  /api/v3/otc/options/{id}/bid:
    post:
      consumes:
      - application/json
      description: Many bidders can each open their own chain against the same listing.
        First to accept wins atomically; siblings cascade-cancel.
      parameters:
      - description: parent OTCOffer listing id
        in: path
        name: id
        required: true
        type: integer
      - description: initial bid terms + bidder's account
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.openNegotiationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "409":
          description: chain already open for caller on this listing
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Place a bid on an OTC option listing (opens a negotiation chain)
      tags:
      - OTCOptions
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"